assignment_1/task_1/grupo_5/final/requirements.txt
boto3
mysql-connector-python
python-dotenv
sqlparse